+++ src/ledgers/vector_results.jl
@@ -55,7 +55,7 @@ const _VECTOR_LANE_TYPES = Dict{Symbol,Type}(
     :b16 => UInt16, :u16 => UInt16, :s16 => Int16,
     :b32 => UInt32, :u32 => UInt32, :s32 => Int32,
     :b64 => UInt64, :u64 => UInt64, :s64 => Int64,
-    :f16 => Float16, :bf16 => BFloat16, :f32 => Float32, :f64 => Float64,
+    :f16 => Float16, :bf16 => UInt16, :f32 => Float32, :f64 => Float64,
     :f16x2 => UInt32, :bf16x2 => UInt32,
     # Scalar FP8 tokens intentionally remain local to this reviewed vector
     # island.  They must not become a generic terminal-result fallback.
